Update to move note id encode/decode function back to model
