+++ lib/models/note.js
@@ -3,6 +3,7 @@
 var fs = require('fs')
 var path = require('path')
 var LZString = require('lz-string')
+var base64url = require('base64url')
 var md = require('markdown-it')()
 var metaMarked = require('meta-marked')
 var cheerio = require('cheerio')
@@ -13,8 +14,6 @@ var moment = require('moment')
 var DiffMatchPatch = require('diff-match-patch')
 var dmp = new DiffMatchPatch()
 var S = require('string')
-
-const {encodeNoteId, decodeNoteId} = require('../utils')
 
 // core
 var config = require('../config')
@@ -116,8 +115,22 @@ module.exports = function (sequelize, DataTypes) {
           return false
         }
       },
-      encodeNoteId: encodeNoteId,
-      decodeNoteId: decodeNoteId,
+      encodeNoteId: function (id) {
+        // remove dashes in UUID and encode in url-safe base64
+        return base64url.encode(id.replace(/-/g, ''))
+      },
+      decodeNoteId: function (encodedId) {
+        // decode from url-safe base64
+        let id = base64url.decode(encodedId)
+        // add dashes between the UUID string parts
+        let idParts = []
+        idParts.push(id.substr(0, 8))
+        idParts.push(id.substr(8, 4))
+        idParts.push(id.substr(12, 4))
+        idParts.push(id.substr(16, 4))
+        idParts.push(id.substr(20, 12))
+        return idParts.join('-')
+      },
       checkNoteIdValid: function (id) {
         var uuidRegex = /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i
         var result = id.match(uuidRegex)
