+++ lib/utils.js
@@ -1,7 +1,6 @@
 'use strict'
 const fs = require('fs')
 const path = require('path')
-const base64url = require('base64url')
 
 exports.isSQLite = function isSQLite (sequelize) {
   return sequelize.options.dialect === 'sqlite'
@@ -32,22 +31,4 @@ exports.isRevealTheme = function isRevealTheme (theme) {
     return theme
   }
   return undefined
-}
-
-exports.encodeNoteId = function (id) {
-  // remove dashes in UUID and encode in url-safe base64
-  return base64url.encode(id.replace(/-/g, ''))
-}
-
-exports.decodeNoteId = function (encodedId) {
-  // decode from url-safe base64
-  let id = base64url.decode(encodedId)
-  // add dashes between the UUID string parts
-  let idParts = []
-  idParts.push(id.substr(0, 8))
-  idParts.push(id.substr(8, 4))
-  idParts.push(id.substr(12, 4))
-  idParts.push(id.substr(16, 4))
-  idParts.push(id.substr(20, 12))
-  return idParts.join('-')
 }
